# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ZEINA Hyperviseur API
  version: "1.0.0"
  description: |
    REST + WebSocket API du superviseur IoT ZEINA.

    ## Authentification
    1. POST `/v1/auth/login` retourne un access token (15 min) + un refresh token (cookie httpOnly, 7 j).
    2. Les routes protégées attendent l'access token dans `Authorization: Bearer <token>`.
    3. Quand le token expire, POST `/v1/auth/refresh` produit un nouvel access (utilise le cookie).

    ## Modèle d'autorisation
    - **Superadmin** (`is_superadmin=true`) : bypass complet.
    - **Owner du tenant** (`tenant_role=owner`) : accès implicite à tous les sites + gestion users/rôles.
    - **Membre simple** : accès aux sites où il est listé dans `site_members`, selon le rôle.
    - Chaque rôle attribue un niveau (`none`/`read`/`write`) sur 4 fonctionnalités : `dashboard`, `devices`, `rules`, `members`.

    ## Test rapide
    1. Clique « Authorize » en haut, paste l'access token (sans le préfixe Bearer).
    2. Tous les calls authentifiés vont passer.
servers:
  - url: /
    description: |
      Origine courante (relatif). Quand tu importes ce spec dans Bruno/Postman,
      remplace par `http://localhost:5173` (via nginx) ou `http://localhost:3000`
      (API directe) — sans slash final.
...
security:
  - bearerAuth: []
paths:
...
  /v1/sites:
    get:
      tags: [Sites]
      summary: Liste des sites accessibles à l'utilisateur courant
      description: Pour un owner/superadmin → tous les sites du tenant. Pour un membre → uniquement les sites où il est dans `site_members`.
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Site" } } } } }
    post:
      tags: [Sites]
      summary: Créer un site (owner / superadmin)
      description: |
        Crée un site et ajoute automatiquement le créateur comme **Responsable de site**
        dans `site_members`. Opération transactionnelle.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSiteReq" }
            examples:
              bobo:
                value:
                  slug: agence-bobo
                  name: Agence Bobo-Dioulasso
                  address: "Rue Verte, Bobo"
                  lat: 11.1771
                  lng: -4.2979
                  timezone: Africa/Ouagadougou
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Site" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/sites/{id}:
    parameters: [{ $ref: "#/components/parameters/SiteId" }]
    get:
      tags: [Sites]
      summary: Détail d'un site
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Site" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
    put:
      tags: [Sites]
      summary: Modifier un site (owner / superadmin)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateSiteReq" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Site" } } } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Sites]
      summary: Supprimer un site (owner / superadmin)
      description: Cascade — supprime aussi zones, devices, dashboards, règles, members et mesures associées.
      responses:
        "204": { description: No content }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/sites/{id}/tree:
    get:
      tags: [Sites]
      summary: Arborescence site → zones → équipements
      parameters: [{ $ref: "#/components/parameters/SiteId" }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/SiteTree" } } } }
  /v1/sites/{id}/summary:
    get:
      tags: [Sites]
      summary: KPI agrégés d'un site (température moyenne, énergie 24 h, etc.)
      parameters: [{ $ref: "#/components/parameters/SiteId" }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/SiteSummary" } } } }

  # ---------------------------------------------------------------------------
  # USERS  (owner ou superadmin)
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SiteId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
      description: ID du site
...
  responses:
...
    Forbidden:
      description: Permission insuffisante
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: Validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error:   { type: string, example: validation }
        message: { type: string, example: "email and password are required" }
        details: { type: object, additionalProperties: true }
...
    Site:
      type: object
      properties:
        id:       { type: string, format: uuid }
        slug:     { type: string }
        name:     { type: string }
        address:  { type: string, nullable: true }
        lat:      { type: number, format: double, nullable: true }
        lng:      { type: number, format: double, nullable: true }
        timezone: { type: string }
    CreateSiteReq:
      type: object
      required: [slug, name]
      properties:
        slug:
          type: string
          pattern: "^[a-z0-9-]+$"
          maxLength: 64
          description: |
            Identifiant pour les topics MQTT (`qlab/<tenant>/<slug>/...`). Minuscules, chiffres, tirets uniquement.
        name:     { type: string }
        address:  { type: string, nullable: true }
        lat:      { type: number, format: double, nullable: true }
        lng:      { type: number, format: double, nullable: true }
        timezone: { type: string, default: "Africa/Ouagadougou" }
    UpdateSiteReq:
      type: object
      properties:
        name:     { type: string, nullable: true }
        address:  { type: string, nullable: true }
        lat:      { type: number, format: double, nullable: true }
        lng:      { type: number, format: double, nullable: true }
        timezone: { type: string, nullable: true }
...
    SiteTree:
      allOf:
        - $ref: "#/components/schemas/Site"
        - type: object
          properties:
            zones:
              type: array
              items:
                type: object
                properties:
                  id:      { type: string, format: uuid }
                  slug:    { type: string }
                  name:    { type: string }
                  devices:
                    type: array
                    items:
                      type: object
                      properties:
                        id:           { type: string, format: uuid }
                        slug:         { type: string }
                        name:         { type: string, nullable: true }
                        type:         { $ref: "#/components/schemas/DeviceType" }
                        status:       { $ref: "#/components/schemas/DeviceStatus" }
                        last_seen_at: { type: string, format: date-time, nullable: true }
    SiteSummary:
      type: object
      properties:
        site_id:             { type: string, format: uuid }
        devices_total:       { type: integer }
        devices_online:      { type: integer }
        energy_day_wh:       { type: number, nullable: true }
        temperature_avg:     { type: number, nullable: true }
        occupancy_ratio_24h: { type: number, nullable: true }

    # --- Users / Roles / Members --------------------------------------------
...
    DeviceType:
      type: string
      enum: [environment, presence, actuator, linky, meter, gateway]
    DeviceStatus:
      type: string
      enum: [provisioned, online, offline, disabled]
...